Adding QuickSort (unfinished) and removing warnings

+++ Algorithms/src/Graph/Components.cpp
@@ -21,7 +21,7 @@ namespace algo
         : m_used(Graph::GetSize(graph))
     {
         size_t graph_size = Graph::GetSize(graph);
-        for (int i = 0; i < graph_size; i++)
+        for (size_t i = 0; i < graph_size; i++)
         {
             if (!m_used[i])
             {
